+++ src/pages/commits/ActiveCommits/CommitChart/CommitChart.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen, userEvent, waitFor } from "test_utils";
 import { ChartTypes } from "types/commits";
-import { CommitChart } from "./CommitChart";
+import { CommitChart } from ".";
 import {
   findMaxGroupedTaskStats,
   getAllTaskStatsGroupedByColor,
-} from "./utils";
+} from "../utils";
 
 describe("commitChart", () => {
   afterEach(() => {
