+++ src/pages/commits/ActiveCommits/CommitChart/CommitChartTooltip.tsx
@@ -6,11 +6,12 @@ import { Disclaimer } from "@leafygreen-ui/typography";
 import { inactiveElementStyle } from "components/styles";
 import { taskStatusToCopy, mapTaskToBarchartColor } from "constants/task";
 import { size, zIndex } from "constants/tokens";
-
+import {
+  ColorCount,
+  getStatusesWithZeroCount,
+} from "pages/commits/ActiveCommits/utils";
 import { TaskStatus } from "types/task";
-
 import { msToDuration } from "utils/string";
-import { ColorCount, getStatusesWithZeroCount } from "./utils";
 
 const { gray } = palette;
 interface Props {
