+++ src/pages/commits/ActiveCommits/buildVariantCard/index.tsx
@@ -5,10 +5,13 @@ import VisibilityContainer from "components/VisibilityContainer";
 import { getVariantHistoryRoute } from "constants/routes";
 import { size } from "constants/tokens";
 import { StatusCount } from "gql/generated/types";
-import { VariantGroupedTaskStatusBadges } from "pages/commits/ActiveCommits/buildVariantCard/VariantGroupedTaskStatusBadges";
-import { TASK_ICON_PADDING } from "../constants";
-import { WaterfallTaskStatusIcon } from "./buildVariantCard/WaterfallTaskStatusIcon";
-import { injectGlobalDimStyle, removeGlobalDimStyle } from "./utils";
+import {
+  injectGlobalDimStyle,
+  removeGlobalDimStyle,
+} from "pages/commits/ActiveCommits/utils";
+import { TASK_ICON_PADDING } from "pages/commits/constants";
+import { VariantGroupedTaskStatusBadges } from "./VariantGroupedTaskStatusBadges";
+import { WaterfallTaskStatusIcon } from "./WaterfallTaskStatusIcon";
 
 type taskList = {
   id: string;
